+++ front-end/src/layouts/loggedin/index.tsx
@@ -2,6 +2,17 @@ import React from "react";
 
 //Components
 import Navbar from "../../components/navbar";
+import Sidebar, { SidebarItem } from "../../components/sidebar";
+
+//Assets
+import {
+	Home,
+	PieChart,
+	Calendar,
+	UserGroup,
+	Users,
+	Edit,
+} from "../../assets/svg";
 
 //Styles
 import "./styles.scss";
@@ -9,7 +20,15 @@ import "./styles.scss";
 function LoggedIn() {
 	return (
 		<div>
-			<Navbar />
+			{/* <Navbar /> */}
+			<Sidebar>
+				<SidebarItem icon={<Home />} text="Dashboard" />
+				<SidebarItem icon={<PieChart />} text="Statistics" active />
+				<SidebarItem icon={<Calendar />} text="Schedules" />
+				<SidebarItem icon={<UserGroup />} text="Visitor Management" />
+				<SidebarItem icon={<Users />} text="Manage Users" />
+				<SidebarItem icon={<Edit />} text="Visitor Home Editor" />
+			</Sidebar>
 		</div>
 	);
 }